Create separate Contex file
